requirements.txt
PyGObject
pillow